requirements.txt
Flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.3.7
python-dotenv==1.0.0
schedule==1.2.0
boto3==1.34.0
elevenlabs==2.5.0
